Refac: move n2v manipulate config to algorithms
