+++ docs/v0.1/careamist_api/configuration/config.yml
@@ -29,7 +29,7 @@ algorithm_config:
     strategy: uniform
     struct_mask_axis: none
     struct_mask_span: 5
-    seed: 395481210
+    seed: 1606660641
 data_config:
   data_type: tiff
   axes: ZYX
