+++ src/careamics/config/algorithms/n2v_manipulation/__init__.py
@@ -1,0 +1,7 @@
+"""N2V manipulation configuration."""
+
+from .n2v_manipulate_config import N2VManipulateConfig
+
+__all__ = [
+    "N2VManipulateConfig",
+]
