+++ src/careamics/config/algorithms/n2v_manipulation/n2v_manipulate_config.py
