+++ src/careamics/config/factories/n2v_factory.py
@@ -4,8 +4,8 @@ from collections.abc import Sequence
 from typing import Any, Literal
 
 from careamics.config.algorithms import N2VAlgorithm
+from careamics.config.algorithms.n2v_manipulation import N2VManipulateConfig
 from careamics.config.augmentations import (
-    N2VManipulateConfig,
     XYFlipConfig,
     XYRandomRotate90Config,
 )
