+++ tests/compat/lightning/test_train_data_module.py
@@ -5,7 +5,8 @@ from tifffile import imwrite
 from careamics.compat.config.data import DataConfig
 from careamics.compat.dataset import InMemoryDataset, PathIterableDataset
 from careamics.compat.lightning import TrainDataModule, create_train_datamodule
-from careamics.config.augmentations import N2VManipulateConfig, XYFlipConfig
+from careamics.config.algorithms.n2v_manipulation import N2VManipulateConfig
+from careamics.config.augmentations import XYFlipConfig
 from careamics.config.support import (
     SupportedData,
     SupportedTransform,
